workaround for Whoosh's handling of __exact queries
By "handling", I mean "ignoring them when they're against a datetime object."
I filed https://github.com/toastdriven/django-haystack/issues/529 for it, but
given that it's only for local development I'm not too hurried about finding a
real fix for it.

It's possible we don't need the `__exact` query at all (just using the
default), in which case this code can be removed.